requirements.txt
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-storages==1.14.4
djangorestframework==3.15.1
boto3==1.35.54
botocore==1.35.54
psycopg2-binary==2.9.9
pillow==10.3.0
requests==2.32.3
urllib3==1.26.18
supabase==1.0.3
tzdata==2024.1
asgiref==3.8.1
anyio==4.3.0
httpx
django-ckeditor


